Extract command line parsing into bin/wavdio-express.js
