+++ .github/workflows/release.yml
@@ -1,6 +1,4 @@
-on:
-  release:
-    types: [ published ]
+on: [push]
 
 jobs:
 
@@ -30,7 +28,7 @@ jobs:
         run: docker run --publish "27017:27017" --detach mongo
 
       - name: Start server
-        run: npx wavdio-express@3.1.0 &
+        run: npx wavdio-express@3.1.0-test &
 
       - name: Wait for server
         uses: cygnetdigital/wait_for_response@v2.0.0
@@ -55,10 +53,10 @@ jobs:
         with:
           push: true
           tags: |
-            keepinmindbot/wavdio-express:latest
-            keepinmindbot/wavdio-express:3
-            keepinmindbot/wavdio-express:3.1
-            keepinmindbot/wavdio-express:3.1.0
+#            keepinmindbot/wavdio-express:latest
+#            keepinmindbot/wavdio-express:3
+#            keepinmindbot/wavdio-express:3.1
+            keepinmindbot/wavdio-express:3.1.0-test
 
   test_docker:
     needs: deliver_docker
@@ -69,7 +67,7 @@ jobs:
         run: |
           docker network create wavdio-net
           docker run --name wavdio-mongo --network wavdio-net --detach mongo
-          docker run --network wavdio-net --publish "3000:3000" --detach keepinmindbot/wavdio-express:3.1.0
+          docker run --network wavdio-net --publish "3000:3000" --detach keepinmindbot/wavdio-express:3.1.0-test
 
       - name: Wait for server
         uses: cygnetdigital/wait_for_response@v2.0.0
