+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "wavdio-express",
-  "version": "3.1.0",
+  "version": "3.1.0-test",
   "scripts": {
-    "start": "node server.js",
+    "start": "node bin/wavdio-express.js",
     "test": "mocha",
     "mocha": "mocha --reporter mocha-junit-reporter --reporter-options mochaFile=./test-results/mocha.xml --exit"
   },
   "bin": {
-    "wavdio-express": "server.js"
+    "wavdio-express": "bin/wavdio-express.js"
   },
   "dependencies": {
     "body-parser": "^1.19.0",
@@ -31,6 +31,6 @@
     "chai-http": "^4.3.0",
     "chai-shallow-deep-equal": "^1.4.6",
     "eslint": "^8.1.0",
-    "mocha": "^9.1.1"
+    "mocha": "^9.1.3"
   }
 }
